+++ view/idioma_lista.php
@@ -1,8 +1,9 @@
 <?php
 session_start();
 
-if(!isset($_SESSION['usu_int_id'])) {
+if(!isset($_SESSION['usu_id'])) {
 	header('location:..\view\erro.php?e=UNL');
+	exit;
 }
 require_once '../lib/biblioteca.php';
 require_once '../model/model.php';
@@ -14,35 +15,35 @@ require_once 'cabecalho.php';
 echo $HEAD_;
 echo $BODY;
 
-require_once 'menu.php';
-
 echo $DIV_MAIN;
-echo $H1 . 'IDIOMA' . $H1_;
+
+require_once 'menu.php';
 
 $FORM = '<form action="../controller/idioma.php" method="get">';
 
 echo $TABLE;
+echo $TR . $TH . 'Idioma' . $TH_ . $TR_;
 
-$idioma_model['usu_int_id'] = $_SESSION['usu_int_id'];
-$condicao = $idioma_model['usu_int_id'];
+$idioma_model['usu_id'] = $_SESSION['usu_id'];
+$condicao = $idioma_model['usu_id'];
 $idioma_json = json_decode(selecionar('idioma', $idioma_model));
 echo $TR . $TD . '<a href="idioma.php">Cadastrar Idioma</a><br>' . $TD_ . $TR_; 
 echo $TR . $TD . $LABEL . '&nbsp;' . $LABEL_ . $TD_ . $TR_;
 
 foreach($idioma_json as $registro) {
 	$idioma_model = array();
-	$idioma_model['idi_int_id'] = $registro->idi_int_id ;	
-	$idioma_model['idi_int_idioma'] = $registro->idi_int_idioma;
-	$idioma_model['idi_int_nivel_conhecimento'] = $registro->idi_int_nivel_conhecimento;
+	$idioma_model['idi_id'] = $registro->idi_id ;	
+	$idioma_model['idi_idioma'] = $registro->idi_idioma;
+	$idioma_model['idi_nivel_conhecimento'] = $registro->idi_nivel_conhecimento;
 	$str = '';
 	foreach ($idioma_model as $k=>$v){ 
 		$str .= "idioma_model[$k]" . "=" . $v . "&";                        
 	}
 
-	echo $TR . $TD . $LABEL . 'Idioma: ' . $array_idioma[$idioma_model['idi_int_idioma']] . $LABEL_ . $TD_ . $TR_; 
-	echo $TR . $TD . $LABEL . 'Nível de conhecimento: ' . $array_nivel_conhecimento_idioma[$idioma_model['idi_int_nivel_conhecimento']] . $LABEL_ . $TD_ . $TR_;  
+	echo $TR . $TD . $LABEL . 'Idioma: ' . $array_idioma[$idioma_model['idi_idioma']] . $LABEL_ . $TD_ . $TR_; 
+	echo $TR . $TD . $LABEL . 'Nível de conhecimento: ' . $array_nivel_conhecimento_idioma[$idioma_model['idi_nivel_conhecimento']] . $LABEL_ . $TD_ . $TR_;  
 	echo $TR . $TD . '<a href="../view/idioma.php?' . $str . '">Editar</a>' . $TD_ . $TR_; 
-	echo $TR . $TD . '<a href="../controller/idioma.php?acao=excluir&' . $str . '">Excluir</a>' . $TD_ . $TR_; 
+	echo $TR . $TD . '<a href="../controller/idioma.php?acao=excluir&' . $str . ' " onclick="return confirmar();">Excluir</a>' . $TD_ . $TR_;  
 	echo $TR . $TD . $LABEL . '&nbsp;' . $LABEL_ . $TD_ . $TR_; 	
 }
